% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr2.R
\name{rrr.control}
\alias{rrr.control}
\title{inner function 1.}
\usage{
rrr.control(sv.tol = 1e-07, qr.tol = 1e-07)
}
\arguments{
\item{sv.tol}{singular value tolerence}

\item{qr.tol}{qr decomposition tolerence}
}
\value{
a list.
}
\description{
null.
}
\examples{
\dontrun{
null
}
}
