% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr2.R
\name{rrr.modstr}
\alias{rrr.modstr}
\title{inner function 2 to give The default value.}
\usage{
rrr.modstr(gamma = 2, nlambda = 100, lambda = NULL)
}
\arguments{
\item{gamma}{inner para with default value 2}

\item{nlambda}{number of lambda}

\item{lambda}{inner para}
}
\value{
a list.
}
\description{
null.
}
\examples{
\dontrun{
null
}
}
