% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_sim6.R
\name{rrr.sim6}
\alias{rrr.sim6}
\title{Auto generated samples.}
\usage{
rrr.sim6(n, p, q, r, b, rho, rX)
}
\arguments{
\item{n}{samples}

\item{p}{number of explanatory variables}

\item{q}{number of response variables}

\item{r}{rank of C}

\item{b}{signal to noise ratio}

\item{rho}{correlation between Xs}

\item{rX}{rank of X}
}
\value{
a list of Y and X and true signal to noise ratio
}
\description{
Simulate according to Chen, Dong, Chan for case where p,q >n.
}
\examples{
\dontrun{
n = 76 # samples
q = 88 # number of response variables
rho = 0.5 # correlation between Xs
rX = 10 # rank of X
r = 5 # rank of C
p = 46 # number of explanatory variables
b = 1 #signal to noise ratio
res <- rrr.sim6(n,p,q,r,b,rho,rX)
}
}
