% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr2.R
\name{softTH}
\alias{softTH}
\title{Soft thresholding}
\usage{
softTH(C, Lam, tol = 1e-04)
}
\arguments{
\item{Lam}{a inner para}

\item{tol}{tolerence}

\item{c}{a inner para}
}
\value{
Soft thresholding
}
\description{
lambda can be a matrix with same dim as C.
}
\examples{
\dontrun{
null
}
}
